requirements.txt
Flask==2.3.2
flask-jwt-extended==4.4.4
flask-pymongo==2.3.0
email-validator==1.3.1
passlib==1.7.4
flask-cors==3.0.10
python-dotenv==1.0.0